enable choice of max-tokens in masked_lm (#1180)
Summary:
# Before submitting

- [ ] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
- [x] Did you read the [contributor guideline](https://github.com/pytorch/fairseq/blob/master/CONTRIBUTING.md)?
- [x] Did you make sure to update the docs?
- [ ] Did you write any new necessary tests?

## What does this PR do?
Adds option to set max-positions in masked_lm model. This option is available in the RoBERTa model, so it should be available here too.

## PR review
Anyone in the community is free to review the PR once the tests have passed.
If we didn't discuss your PR in Github issues there's a high chance it will not be merged.

## Did you have fun?
Make sure you had fun coding �
Pull Request resolved: https://github.com/fairinternal/fairseq-py/pull/1180

Reviewed By: myleott

Differential Revision: D21904170

Pulled By: joshim5

fbshipit-source-id: 37168dbf1a2758620d5e8e05c7e8a9ef8d09c765