Add tracepoints (#1192)
Summary:
There is no overhead when the profiling is not enabled.
When running using profile.py, I measure an overhead of 3%.

# Before submitting

- [ ] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
- [ ] Did you read the [contributor guideline](https://github.com/pytorch/fairseq/blob/master/CONTRIBUTING.md)?
- [ ] Did you make sure to update the docs?
- [ ] Did you write any new necessary tests?

## What does this PR do?
Fixes # (issue).

## PR review
Anyone in the community is free to review the PR once the tests have passed.
If we didn't discuss your PR in Github issues there's a high chance it will not be merged.

## Did you have fun?
Make sure you had fun coding �
Pull Request resolved: https://github.com/fairinternal/fairseq-py/pull/1192

Reviewed By: sidgoyal78

Differential Revision: D22102341

Pulled By: msbaines

fbshipit-source-id: ffddb9cceb853df88db34195be18bae7723d4c98