Fix memory leak with small data-buffer-size
Summary:
As part of zhengwy888's debugging of a memory leak, he suggested that trimming the number of batches in pyspeech's train.py may cause the BufferedIterator to leave some batches in the queue, causing a memory leak.

Therefore, propagating `take` to the buffered iterator, which should prevent the consumer thread from hanging on `queue.put`.

Reviewed By: myleott

Differential Revision: D22405263

fbshipit-source-id: 80f40a355652016af4ba8c386b623cb0552b1928