Replace encoder output type (#1281)
Summary:
Pull Request resolved: https://github.com/fairinternal/fairseq-py/pull/1281

The PyTorch Mobile lite interpreter does not support NamedTuple creation in forward. One workaround is to replace NamedTuple with a custom class that inherits nn.Module. This class could be initialized in `__init__` and updated in forward. However, lite interpreter does not support list construction with custom classes. So the final solution is to replace the NamedTuple with a dictionary. We cannot have mixed value types in that dictionary, otherwise, this breaks TorchScript export. So the type is List[Tensor] and an empty list corresponds to having a value of None.

Reviewed By: myleott

Differential Revision: D23752010

fbshipit-source-id: 0b152a534a165ce4f84bd4f580d7f29145cfd264