More accurate label smoothing loss computation (#3182)
Summary:
It seems that the current implementation uses a slightly larger label smoothing value, for a large vocabulary, it is fine, but it can be more different with a small vocabulary size. By changing these 2 lines, the computation of label smoothing loss shall be consistent with the standard.

Pull Request resolved: https://github.com/pytorch/fairseq/pull/3182

Reviewed By: glample

Differential Revision: D26225506

Pulled By: myleott

fbshipit-source-id: 75447275e32336ae3b52e732e6124e15d0043b74