sorting os.listdir outputs to ensure consistent import ordering
Summary:
There's a common idiom of doing imports based on the output of `os.listdir(os.path.dirname(__file__))`.  In this idiom, imports are performed based on the order of directories in that output.

This is dangerous, because the behavior of your program depends on the order of those files, but the order of those files is not guaranteed.

Reviewed By: zhengwy888

Differential Revision: D27951383

fbshipit-source-id: 97dc9a0b7d853886e19a9643c33508f146c71617