Fixes fp16 inference for speech recognition example (#2438)
Summary:
Currently the speech recognition fails for fp16 option as the input is
not converted to half. This fixes this issue.

# Before submitting

- [ ] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
- [x] Did you read the [contributor guideline](https://github.com/pytorch/fairseq/blob/main/CONTRIBUTING.md)?
- [x] Did you make sure to update the docs?
- [ ] Did you write any new necessary tests?

## What does this PR do?
Fixes fp16 inference for speech recognition example.

## PR review
Anyone in the community is free to review the PR once the tests have passed.
If we didn't discuss your PR in Github issues there's a high chance it will not be merged.

## Did you have fun?
Make sure you had fun coding �

Pull Request resolved: https://github.com/fairinternal/fairseq-py/pull/2438

Reviewed By: apoorv2904, xuqiantong

Differential Revision: D31594529

Pulled By: alexeib

fbshipit-source-id: 75d1fe3f00b2e86381f28f79daea2cd516233e22